Consolidate translateBufferLineToString functions
