+++ src/SearchHelper.ts
@@ -1,4 +1,9 @@
+/**
+ * @license MIT
+ */
+
 import { ITerminal } from './Interfaces';
+import { translateBufferLineToString } from './utils/BufferLine';
 
 interface ISearchResult {
   term: string;
@@ -98,7 +103,7 @@ export class SearchHelper {
 
   private _findInLine(term: string, y: number): ISearchResult {
     const bufferLine = this._terminal.lines.get(y);
-    const stringLine = this._translateBufferLineToString(bufferLine, true);
+    const stringLine = translateBufferLineToString(bufferLine, true);
     const searchIndex = stringLine.indexOf(term);
     if (searchIndex >= 0) {
       return {
@@ -116,49 +121,5 @@ export class SearchHelper {
     this._terminal.selectionManager.setSelection(result.col, result.row, result.term.length);
     this._terminal.scrollDisp(result.row - this._terminal.ydisp, false);
     return true;
-  }
-
-  // TODO: Consolidate with SelectionManager function
-  private _translateBufferLineToString(line: any, trimRight: boolean, startCol: number = 0, endCol: number = null): string {
-    // TODO: This function should live in a buffer or buffer line class
-
-    // TODO: Move these constants elsewhere, they belong in a buffer or buffer
-    //       data/line class.
-    const LINE_DATA_CHAR_INDEX = 1;
-    const LINE_DATA_WIDTH_INDEX = 2;
-    // Get full line
-    let lineString = '';
-    let widthAdjustedStartCol = startCol;
-    let widthAdjustedEndCol = endCol;
-    for (let i = 0; i < line.length; i++) {
-      const char = line[i];
-      lineString += char[LINE_DATA_CHAR_INDEX];
-      // Adjust start and end cols for wide characters if they affect their
-      // column indexes
-      if (char[LINE_DATA_WIDTH_INDEX] === 0) {
-        if (startCol >= i) {
-          widthAdjustedStartCol--;
-        }
-        if (endCol >= i) {
-          widthAdjustedEndCol--;
-        }
-      }
-    }
-
-    // Calculate the final end col by trimming whitespace on the right of the
-    // line if needed.
-    let finalEndCol = widthAdjustedEndCol || line.length;
-    if (trimRight) {
-      const rightWhitespaceIndex = lineString.search(/\s+$/);
-      if (rightWhitespaceIndex !== -1) {
-        finalEndCol = Math.min(finalEndCol, rightWhitespaceIndex);
-      }
-      // Return the empty string if only trimmed whitespace is selected
-      if (finalEndCol <= widthAdjustedStartCol) {
-        return '';
-      }
-    }
-
-    return lineString.substring(widthAdjustedStartCol, finalEndCol);
   }
 }
